Collection element new picture addition refact ok
