+++ server/middleware/multerConfig.js
@@ -1,6 +1,9 @@
 const multer=require('multer')
 require('dotenv').config()
+
+
 const collectionPath=process.env.COLLECTIONPICTUREPATH
+const collectionElementPath=process.env.COLLECTIONELEMENTPICTURESPATH
 
 const storage=multer.diskStorage({
     destination: function(req, file, cb){
@@ -14,4 +17,19 @@ const storage=multer.diskStorage({
 
 const upload=multer({storage:storage})
 
-module.exports=upload
+
+
+
+const storageCollectionElementPicture=multer.diskStorage({
+    destination: function(req, file, cb){
+        cb(null, collectionElementPath);
+    },
+    filename:function(req, file, cb){
+        cb(null, file.originalname)
+    }
+})
+
+
+const uploadCollectionElementPicture=multer({storage:storageCollectionElementPicture})
+
+module.exports={upload, uploadCollectionElementPicture}
