+++ server/server.js
@@ -30,104 +30,6 @@ app.use((req, res, next) => {
 
 
 
-//Create Collection Element informations
-
-
-app.post('/admin/createCollectionElementInformations',async(req, res,next)=>{
-    try {
-        console.log('informations to Create => ',req.body)
-        let {informationsToCreate,collectionUID}=req.body
-
-        await pool.query('BEGIN')
-        if (informationsToCreate.length>0){
-
-            for (let informationToCreate of informationsToCreate){
-                const {informationInputText}=informationToCreate
-            let collectionElementInformationToCreate=await pool.query(
-                `INSERT INTO collection_element_informations
-                (collection_uid, collection_element_information_text)
-                VALUES ($1, $2)
-                `,[collectionUID,informationInputText]
-            )
-            if(collectionElementInformationToCreate.rowCount===0){
-                console.log(`collection information was NOT created for collection_uid ${collectionUID}`)
-            }
-            else console.log(`collection information was created for collection_uid ${collectionUID}`)
-            }
-
-            await pool.query('COMMIT')
-            res.status(200).json({message: 'collection element informations were created'})
-        }
-        else console.log('No information was received')
-
-    } catch (error) {
-        await pool.query('ROLLBACK')
-        console.error(`error ceating collection element informations=> ${error} `)
-        return res.status(400).json({message:"Creation for collection Element informations wasn't completed due to an error"})
-    }
-
-})
-
-
-
-
-//Post 1 new Collection Element Picture
-
-
-const storageCollectionElementPicture=multer.diskStorage({
-    destination: function(req, file, cb){
-        cb(null, collectionElementPath);
-    },
-    filename:function(req, file, cb){
-        cb(null, file.originalname)
-    }
-})
-
-
-const uploadCollectionElementPicture=multer({storage:storageCollectionElementPicture})
-
-app.post('/admin/editElement/addNewPicture', uploadCollectionElementPicture.single('file'), async(req, res, next)=>{
-    const {collectionUID}=req.body
-    try{
-        const collectionName=await pool.query(
-            `SELECT collection_title FROM collection
-            WHERE collection_uid=$1`,[collectionUID]
-        )
-        if (collectionName){
-            const collectionDirectory=collectionName.rows[0].collection_title
-            console.log('collection Directory =>',collectionDirectory)
-            const collectionElementPictureAlt=`Image ${req.file.originalname}`
-            const collectionElementPictureUrl=`/images/${req.file.originalname}`
-            
-            try{
-                const newCollectionElementPicture=await pool.query(
-                    'INSERT INTO collection_element_pictures (collection_uid, collection_element_picture_url, collection_element_picture_alt, collection_element_pictures_deletionflag) VALUES ($1, $2, $3, $4) RETURNING *',
-                    [collectionUID, collectionElementPictureUrl, collectionElementPictureAlt, false]
-                )
-                if(newCollectionElementPicture){
-                    return res.status(200).json({message:newCollectionElementPicture.rows[0]})
-                }
-                else {
-                    return res.status(201).json({message:"picture not added"})
-                }
-            }
-            catch(error){
-                console.error('Error adding newCollectionElemenPicture', error)
-                return res.status(500).send('Server Error')
-            }
-        }
-    }
-    catch (error){
-        console.error('Error getting Collection UID', error)
-        return res.status(500).send('Server Error 2')
-
-    }
-
-})
-
-
-
-
 //Delete 1 collection
